Refactor LEDMatrixController structure; add new components for keep-alive, animation, drawing, identification, brightness, and breathing management
