+++ is_matrix_forge/led_matrix/controller/__init__.py
@@ -1,27 +1,16 @@
-"""
-LED Matrix Controller Module.
-
-This module provides the main controller class for interacting with LED matrix devices.
-It handles device communication, animation control, pattern display, and other
-operations related to the LED matrix hardware.
-"""
+"""LED Matrix Controller module."""
 
 from __future__ import annotations
 
-# Standard library
 from typing import Optional
 
-# Third‑party
-
-# Inspyre‑Softworks
-
-from .multiton import MultitonMeta
 from is_matrix_forge.led_matrix.constants import HEIGHT, WIDTH
-
 from is_matrix_forge.led_matrix.display.grid.grid import Grid
-from .helpers.threading import synchronized
+
 from .controller import LEDMatrixController
 from .helpers import get_controllers
+from .multiton import MultitonMeta
+from .helpers.threading import synchronized
 
 
 def generate_blank_grid(width: Optional[int] = None, height: Optional[int] = None) -> Grid:
